+++ src/main/java/Item/Repair/FirearmRepairKit.java
@@ -30,8 +30,9 @@ public class FirearmRepairKit implements IScriptEventHandler {
             repairAmount *= 1.25f;
         }
 
-        DurabilitySystem.RunItemRepair(oPC, item, repairAmount);
-
-        NWScript.destroyObject(item, 0.0f);
+        if(DurabilitySystem.RunItemRepair(oPC, target, repairAmount))
+        {
+            NWScript.destroyObject(item, 0.0f);
+        }
     }
 }
